Renamed utils.logging to utils.log because of clashing name with logging package
