+++ naslib/optimizers/discrete/npenas/optimizer.py
@@ -17,7 +17,7 @@ from naslib.predictors.utils.encodings import encode_spec
 from naslib.search_spaces.core.query_metrics import Metric
 
 from naslib.utils import AttrDict, count_parameters_in_MB, get_train_val_loaders
-from naslib.utils.logging import log_every_n_seconds
+from naslib.utils.log import log_every_n_seconds
 
 
 logger = logging.getLogger(__name__)
